Completed fort.15 generator that works for tidal runs.
This commit completes the fort.15 generation for the
AdcircMesh.generate_tidal_run() method. When called with the proper
arguments, this method generates two files: fort.15.coldstart and
fort.15.hotstart.

It was decided that this package will always divide the coldstart and
hotstart phases, even on tidal only runs. That way the modeler can keep
the output files separate. By default, the coldstart phase will not
generate outputs, and will only generate a single hotstart file at the
end of the spinup. This file can then be used in tandem with the
fort.15.hotstart in order to generate the actual simulation run.

Stations oputputs and global outputs may be passed as arguments to the
generate_tidal_run() method in order to request outputs. See
tests/test_generate_90dayTidal.py to see an example usage of output
requests.

This commit also  includes minor fixes to the matplotlib calls on the
test files.