Best track files (fort.22) from HURDAT2 database
This commit faithfully reproduces the fort.22 present in NOAA's official
run for Sandy with NWS=20. The only difference is that HURDAT2 is more
accurate in timing of the landfall than HURDAT1. Not yet tested with
ADCIRC. Also implementation of radius of maximum winds is missing, and
implementations for creating perturbations is also in the to-do list.