Reintroduced DEM tools
Many operations require the manipulation of DEMs. Although some have
suggested to have GDAL removed, I believe GDAL is an important part of
Geospatial data manipulation and thus one shall not "throw out the baby
with the bath water". This DEM module in reality requires more work, but
there are functions here that are useful and should be kept and improved
upon.