Changed the way parse_fort14() behaves.
The new behaviour assumes EOF is NOPE line is empty. Before it would
expect NOPE to be an int (0 for no boundaries, instead of empty).
Accomodated the ability to read in NOPE=-1 to indicate self-closing
mesh.

This is a bugfix of the previous commit, where this update had been
applied to the NBOU variable instead of the NOPE variable.